FastAPI/Alugaphone/aluga-project/pyproject.toml
[tool.poetry]
name = "aluga-project"
version = "0.1.0"
description = ""
authors = ["PedroElorriaga <pedrohes2002@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = "^2.0.30"
pydantic-settings = "^2.2.1"
alembic = "^1.13.1"
factory-boy = "^3.3.0"
pyjwt = "^2.8.0"
pwdlib = {extras = ["argon2"], version = "^0.2.0"}
python-multipart = "^0.0.9"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-cov = "^5.0.0"
taskipy = "^1.12.2"
blue = "^0.9.1"
ruff = "^0.4.4"
httpx = "^0.27.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
line-length = 79
exclude = ['.venv', 'migrations']


[tool.isort]
profile = "black"
line_length = 79
extend_skip = ['migrations']


[tool.pytest.ini_options]
pythonpath = "."


[tool.blue]
extend-exclude = '(migrations/)'


[tool.taskipy.tasks]
lint = 'ruff . && blue --check . --diff && isort --check . --diff'
format = 'blue .  && isort .'
run = 'uvicorn aluga_project.app:app --reload'
pre_test = 'task lint'
test = 'pytest -s -x --cov=aluga_project -vv'
post_test = 'coverage html'